+++ src/WireMock.Net/RequestBuilders/Request.WithBody.cs
@@ -2,8 +2,6 @@
 // For more details see 'mock4net/LICENSE.txt' and 'mock4net/readme.md' in this project root.
 using System;
 using System.Collections.Generic;
-using JsonConverter.Abstractions;
-using Newtonsoft.Json;
 using Stef.Validation;
 using WireMock.Matchers;
 using WireMock.Matchers.Request;
@@ -37,19 +35,7 @@ public partial class Request
     /// <inheritdoc />
     public IRequestBuilder WithBodyAsJson(object body, MatchBehaviour matchBehaviour = MatchBehaviour.AcceptOnMatch)
     {
-        var bodyAsJsonString = JsonConvert.SerializeObject(body);
-        _requestMatchers.Add(new RequestMessageBodyMatcher(matchBehaviour, bodyAsJsonString));
-        return this;
-    }
-
-    /// <inheritdoc />
-    public IRequestBuilder WithBodyAsJson(object body, IJsonConverter converter, JsonConverterOptions? options = null, MatchBehaviour matchBehaviour = MatchBehaviour.AcceptOnMatch)
-    {
-        Guard.NotNull(converter);
-
-        var bodyAsJsonString = converter.Serialize(body, options);
-        _requestMatchers.Add(new RequestMessageBodyMatcher(matchBehaviour, bodyAsJsonString));
-        return this;
+        return WithBody(new IMatcher[] { new JsonMatcher(matchBehaviour, body) });
     }
 
     /// <inheritdoc />